fix: edit handler & arttoy file name
